+++ lib/EpiHouseholds/test/runtests.jl
@@ -853,5 +853,6 @@ end
 include("test_contextual_kernels.jl")
 include("test_likelihood_composition.jl")
 include("test_initial_cases.jl")
+include("test_actions.jl")
 
 include("test_calendar_kernels.jl")